pyproject.toml
[tool.poetry]
name = "xbar-google-calendar"
version = "0.1.0"
description = ""
authors = ["alex <alex@pdp7.net>"]

[tool.poetry.dependencies]
python = "~3.9"
google-api-python-client = "^2.97.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
